LayoutTests/imported/w3c/IndexedDB-private-browsing/idbcursor_continue_index4-expected.txt


PASS IDBCursor.continue() - index - attempt to iterate to the next record when the direction is set for the previous record
